Support encrypted secrets in lists
With this change, we allow encrypted secrets in lists like

    users:
      - username: Alice
        password: !vault |
          $ANSIBLE_VAULT;1.1;AES256
          363062663635....
      - username: Bob
        password: !vault |
          $ANSIBLE_VAULT;1.1;AES256
          336665396534....

Before this change, Spring didn't properly create the user objects because
the passwords and the usernames are stored in different property sources.
Either the username or the password of each user was null. It is now fixed
because the username and the decrypted password are stored in the same
property source.

Co-authored-by: Thorsten Nuss <thorsten.nuss@gmail.com>